LPD-37074 Move dynamic-data-mapping-form-web files to default environment subfolder
