+++ modules/test/playwright/tests/dynamic-data-mapping-form-web/default-environment/dependencies/form-with-data-provider.portlet.lar/group/20117/com.liferay.dynamic.data.mapping.model.DDMStructure/35153/ddm-form-layout.json
