+++ modules/test/playwright/tests/dynamic-data-mapping-form-web/default-environment/dependencies/form-with-data-provider.portlet.lar/group/20117/portlet/com_liferay_dynamic_data_mapping_form_web_portlet_DDMFormAdminPortlet/20117/portlet-data.xml
