+++ modules/test/playwright/tests/dynamic-data-mapping-form-web/default-environment/dependencies/form-with-data-provider.portlet.lar/manifest.xml
