+++ modules/test/playwright/tests/dynamic-data-mapping-form-web/default-environment/richText.spec.ts
@@ -5,11 +5,11 @@
 
 import {Page, expect, mergeTests} from '@playwright/test';
 
-import {featureFlagsTest} from '../../fixtures/featureFlagsTest';
-import {formsPagesTest} from '../../fixtures/formsPagesTest';
-import {loginTest} from '../../fixtures/loginTest';
-import {getRandomInt} from '../../utils/getRandomInt';
-import {deleteItems} from './utils/deleteItems';
+import {featureFlagsTest} from '../../../fixtures/featureFlagsTest';
+import {formsPagesTest} from '../../../fixtures/formsPagesTest';
+import {loginTest} from '../../../fixtures/loginTest';
+import {getRandomInt} from '../../../utils/getRandomInt';
+import {deleteItems} from '../utils/deleteItems';
 
 export const xssBypassTest = mergeTests(
 	featureFlagsTest({
